feature: implement paralel routes
